requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
sentence-transformers
faiss-cpu
chromadb
langchain
fastapi
uvicorn
streamlit
jupyterlab
ipykernel
python-dotenv
faiss-cpu
transformers
accelerate
tabulate
